+++ gridopt/power_flow/dc_opf.py
@@ -6,11 +6,11 @@
 # GRIDOPT is released under the BSD 2-clause license. #
 #*****************************************************#
 
+from __future__ import print_function
 import numpy as np
 from .method_error import *
 from .method import PFmethod
 from numpy.linalg import norm
-from scipy.sparse import triu,coo_matrix,bmat,eye
 
 class DCOPF(PFmethod):
     """
@@ -19,20 +19,34 @@ class DCOPF(PFmethod):
 
     name = 'DCOPF'
         
-    parameters = {'quiet' : False,
-                  'thermal_limits': True,
-                  'thermal_factor': 1.,
-                  'inf_flow': 1e4,
-                  'vargen_curtailment': False}
-                                    
+    parameters = {'thermal_limits': False,
+                  'renewable_curtailment': False,
+                  'optsolver' : 'iqp'}
+
+    parameters_iqp = {}
+
+    parameters_augl = {}
+
+    parameters_ipopt = {}
+    
     def __init__(self):
 
-        from optalg.opt_solver import OptSolverIQP
+        from optalg.opt_solver import OptSolverIQP, OptSolverAugL, OptSolverIpopt
 
+        # Parent init
         PFmethod.__init__(self)
-        parameters = OptSolverIQP.parameters.copy()
-        parameters.update(DCOPF.parameters)
-        self.parameters = parameters
+        
+        # Optsolver parameters
+        iqp_params = OptSolverIQP.parameters.copy()
+        iqp_params.update(self.parameters_iqp)     # overwrite defaults
+        augl_params = OptSolverAugL.parameters.copy()
+        augl_params.update(self.parameters_augl)   # overwrite defaults
+        ipopt_params = OptSolverIpopt.parameters.copy()
+        ipopt_params.update(self.parameters_ipopt) # overwrite defaults
+        self.parameters.update(DCOPF.parameters)
+        self.parameters['optsolver_params'] = {'iqp': iqp_params,
+                                               'augl': augl_params,
+                                               'ipopt': ipopt_params}
 
     def create_problem(self,net):
 
@@ -40,6 +54,7 @@ class DCOPF(PFmethod):
         
         # Parameters
         params = self.parameters
+        thermal_limits = params['thermal_limits']
         
         # Clear flags
         net.clear_flags()
@@ -57,7 +72,7 @@ class DCOPF(PFmethod):
                       ['variable','bounded'],
                       'adjustable active power',
                       'active power')
-        if params['vargen_curtailment']:
+        if params['renewable_curtailment']:
             net.set_flags('variable generator',
                           ['variable','bounded'],
                           'any',
@@ -66,7 +81,7 @@ class DCOPF(PFmethod):
         try:
             num_gvar =  len([g for g in net.generators if 
                              (not g.is_on_outage()) and g.is_P_adjustable()])
-            num_cur = net.num_var_generators if params['vargen_curtailment'] else 0
+            num_cur = net.num_var_generators if params['renewable_curtailment'] else 0
             assert(net.num_bounded == (num_gvar+net.get_num_P_adjust_loads()+num_cur)*net.num_periods)
             assert(net.num_vars == (net.num_buses-net.get_num_slack_buses()+
                                     num_gvar+net.get_num_P_adjust_loads()+
@@ -78,7 +93,8 @@ class DCOPF(PFmethod):
         problem = pfnet.Problem(net)
         problem.add_constraint(pfnet.Constraint('variable bounds',net))
         problem.add_constraint(pfnet.Constraint('DC power balance',net))
-        problem.add_constraint(pfnet.Constraint('DC branch flow limits',net))
+        if thermal_limits:
+            problem.add_constraint(pfnet.Constraint('DC branch flow limits',net))
         problem.add_function(pfnet.Function('generation cost',1.,net))
         problem.add_function(pfnet.Function('consumption utility',-1.,net))
         problem.analyze()
@@ -88,96 +104,27 @@ class DCOPF(PFmethod):
             
     def solve(self,net):
 
-        from optalg.opt_solver import OptSolverError,OptSolverIQP,QuadProblem
+        from optalg.opt_solver import OptSolverError
+        from optalg.opt_solver import OptSolverIQP, OptSolverAugL, OptSolverIpopt
         
         # Parameters
         params = self.parameters
-        thermal_limits = params['thermal_limits']
-        thermal_factor = params['thermal_factor']
-        inf_flow = params['inf_flow']
+        optsolver_name = params['optsolver']
+        optsolver_params = params['optsolver_params']
 
-        # Check parameters
-        if thermal_factor < 0.:
-            raise PFmethodError_BadParam(self,'thermal_factor')
+        # Opt solver
+        if optsolver_name == 'iqp':
+            optsolver = OptSolverIQP()
+        elif optsolver_name == 'augl':
+            optsolver = OptSolverAugL()
+        elif optsolver_name == 'ipopt':
+            optsolver = OptSolverIpopt()
+        else:
+            raise PFmethodError_BadOptSolver(self)
+        optsolver.set_parameters(optsolver_params[optsolver_name])
 
         # Problem
         problem = self.create_problem(net)
-
-        # Renewables
-        Pr = net.get_var_projection('variable generator','active power')
-       
-        # Construct QP
-        x = problem.get_init_point()
-        problem.eval(x)
-        c_flows = problem.find_constraint('DC branch flow limits')
-        c_bounds = problem.find_constraint('variable bounds')
-        
-        Hx = problem.Hphi + problem.Hphi.T - triu(problem.Hphi)
-        gx = problem.gphi - Hx*x
-        
-        Ax = problem.A
-        bx = problem.b
-        
-        lz = c_flows.l
-        uz = c_flows.u
-        Gz = c_flows.G
-
-        # Flow limit expansion
-        dz = (thermal_factor-1.)*(uz-lz)/2.
-        if not thermal_limits:
-            dz += inf_flow
-        lz -= dz
-        uz += dz
-  
-        lx = c_bounds.l
-        ux = c_bounds.u
-        Gx = c_bounds.G
-        
-        ux += Pr.T*Pr*(x-ux) # correct limit for curtailment
-
-        nx = net.num_vars
-        nz = Gz.shape[0]
-        n = nx+nz
-
-        Iz = eye(nz)
-        Oz = coo_matrix((nz,nz))
-        oz = np.zeros(nz)
-
-        H = bmat([[Hx,None],[None,Oz]],format='coo')
-        g = np.hstack((gx,oz))
-
-        A = bmat([[Ax,None],[Gz,-Iz]],format='coo')
-        b = np.hstack((bx,oz))
-
-        l = np.hstack((lx,lz))
-        u = np.hstack((ux,uz))
-
-        y = np.hstack((x,oz))
-
-        # Check flow limits
-        if not np.all(lz < uz):
-            raise PFmethodError_BadFlowLimits(self)
-        
-        # Check variable limits
-        if not np.all(lx < ux):
-            raise PFmethodError_BadVarLimits(self)
-
-        # Other checks
-        try:
-            assert(Gx.shape == (nx,nx))
-            assert(np.all(Gx.row == Gx.col))
-            assert(np.all(Gx.data == np.ones(nx)))
-            assert(Gz.shape == (len([br for br in net.branches if br.ratingA != 0.])*net.num_periods,nx))
-            assert(l.shape == (n,))
-            assert(u.shape == (n,))
-            assert(np.all(l < u))
-            assert(norm(np.hstack((problem.gphi,oz))-(H*y+g)) < 1e-10*(1.+norm(problem.gphi)))
-            assert(H.shape == (n,n))
-            assert(A.shape == (net.num_buses*net.num_periods+nz,n))
-        except AssertionError:
-            raise PFmethodError_BadProblem(self)
-            
-        QPproblem = QuadProblem(H,g,A,b,l,u)
         
         # Set up solver
         solver = OptSolverIQP()
@@ -185,20 +132,20 @@ class DCOPF(PFmethod):
         
         # Solve
         try:
-            solver.solve(QPproblem)
+            optsolver.solve(problem)
         except OptSolverError as e:
             raise PFmethodError_SolverError(self,e)
         finally:
 
             # Update net properties
-            net.update_properties(solver.get_primal_variables()[:net.num_vars])
+            net.update_properties(optsolver.get_primal_variables()[:net.num_vars])
             
             # Get results
-            self.set_status(solver.get_status())
-            self.set_error_msg(solver.get_error_msg())
-            self.set_iterations(solver.get_iterations())
-            self.set_primal_variables(solver.get_primal_variables())
-            self.set_dual_variables(solver.get_dual_variables())
+            self.set_status(optsolver.get_status())
+            self.set_error_msg(optsolver.get_error_msg())
+            self.set_iterations(optsolver.get_iterations())
+            self.set_primal_variables(optsolver.get_primal_variables())
+            self.set_dual_variables(optsolver.get_dual_variables())
             self.set_net_properties(net.get_properties())
             self.set_problem(problem)
 
@@ -209,35 +156,27 @@ class DCOPF(PFmethod):
     
         # Get data
         problem = self.results['problem']
-        xz = self.results['primal variables']
+        x = self.results['primal variables']
         lam,nu,mu,pi = self.results['dual variables']
-        nx = net.num_vars
-        nz = len([br for br in net.branches if br.ratingA != 0.])*net.num_periods
-        n = nx+nz
-        
+       
         # No problem
         if problem is None:
             raise PFmethodError_NoProblem(self)
- 
+
         # Checks
-        assert(problem.x.shape == (nx,))
-        assert(problem.A.shape == (net.num_buses*net.num_periods,nx))
-        assert(problem.G.shape == (nx+nz,nx))
-        assert(xz.shape == (nx+nz,))
-        assert(lam.shape == (net.num_buses*net.num_periods+nz,))
-        assert(mu.shape == (n,))
-        assert(pi.shape == (n,))
-        assert(nu is None or not nu.size)
+        assert(problem.x.shape == x.shape)
+        assert(net.num_vars+problem.num_extra_vars == x.size)
+        assert(problem.A.shape[0] == lam.size)
+        assert(problem.f.shape[0] == nu.size)
+        assert(problem.G.shape[0] == mu.size)
+        assert(problem.G.shape[0] == pi.size)
 
         # Network quantities
-        net.set_var_values(xz[:nx])
+        net.set_var_values(x[:net.num_vars])
 
         # Network properties
         net.update_properties()
         
         # Network sensitivities
         net.clear_sensitivities()
-        problem.store_sensitivities(lam[:net.num_buses*net.num_periods],
-                                    nu,
-                                    mu,
-                                    pi)
+        problem.store_sensitivities(lam,nu,mu,pi)