+++ tests/test_power_flow.py
@@ -32,9 +32,10 @@ class TestPowerFlow(unittest.TestCase):
         
         for case in utils.test_cases:
             for sol in list(sol_types.keys()):
-                for method_name in ['NRPF','AugLPF']:
+                for optsolver in ['nr','augl','ipopt', 'inlp']:
                     
-                    method = gopt.power_flow.new_method(method_name)
+                    method = gopt.power_flow.new_method('ACPF')
+                    method.set_parameters(params={'optsolver':optsolver})
                     
                     net = pf.Parser(case).parse(case)
                     netMP = pf.Parser(case).parse(case,T)
@@ -48,16 +49,16 @@ class TestPowerFlow(unittest.TestCase):
                     # No sol
                     if sol_data is None:
                         continue
-                    
-                    # Skip
-                    if (sol == 'sol1' and method_name == 'AugLPF'):
+                                          
+                    # Skip (no controls is not supported by augl,ipopt formulation
+                    if sol == 'sol1' and optsolver in ['augl','ipopt','inlp']:
                         continue
 
                     # Set parameters
                     if sol == 'sol1':
                         method.set_parameters({'limit_gens': False})
                     elif sol == 'sol2':
-                        pass
+                        pass # defaults
                     elif sol == 'sol3':
                         method.set_parameters({'lock_taps': False})
                     else:
@@ -65,7 +66,10 @@ class TestPowerFlow(unittest.TestCase):
                     method.set_parameters({'quiet': True})
 
                     bus_P_mis = net.bus_P_mis
-                    method.solve(net)
+                    try:
+                        method.solve(net)
+                    except ImportError:
+                        continue # no ipopt
                     results = method.get_results()
                     self.assertEqual(results['status'],'solved')
                     self.assertEqual(net.bus_P_mis,bus_P_mis)
@@ -111,7 +115,10 @@ class TestPowerFlow(unittest.TestCase):
                     if len(v_mag_error) > 0:
                         self.assertLessEqual(np.max(v_mag_error),v_mag_tol)
                         self.assertLessEqual(np.max(v_ang_error),v_ang_tol)
-                        print("%s\t%s\t%s" %(method_name,case.split('/')[-1],sol_types[sol]))
+                        print("%s\t%s\t%s\t%d" %(case.split('/')[-1],
+                                                 sol_types[sol],
+                                                 optsolver,
+                                                 results['iterations']))
                     self.assertEqual(len(v_mag_error),counter*(T+1))
                     self.assertEqual(len(v_ang_error),counter*(T+1))
 
@@ -121,10 +128,14 @@ class TestPowerFlow(unittest.TestCase):
         
         eps = 0.5 # %
         
-        method_ipopt = gopt.power_flow.new_method('IpoptOPF')
-        method_augl = gopt.power_flow.new_method('AugLOPF')
+        method_ipopt = gopt.power_flow.new_method('ACOPF')
+        method_ipopt.set_parameters(params={'optsolver':'ipopt','quiet': True})
+        method_augl = gopt.power_flow.new_method('ACOPF')
+        method_augl.set_parameters(params={'optsolver':'augl','quiet': True})
+        method_inlp = gopt.power_flow.new_method('ACOPF')
+        method_inlp.set_parameters(params={'optsolver':'inlp','quiet': True})
 
-        skipcases = ['aesoSL2014.raw','case9241pegase.mat','case32.art']
+        skipcases = ['aesoSL2014.raw','case3012wp.mat','case9241.mat','case32.art']
             
         for case in utils.test_cases:
 
@@ -134,10 +145,8 @@ class TestPowerFlow(unittest.TestCase):
             net = pf.Parser(case).parse(case)
             
             self.assertEqual(net.num_periods,1)
-            
-            method_ipopt.set_parameters({'quiet':True})
-            method_augl.set_parameters({'quiet':True})
 
+            # IPOPT
             try:
                 net.update_properties()
                 gen_P_cost = net.gen_P_cost
@@ -147,35 +156,53 @@ class TestPowerFlow(unittest.TestCase):
                 self.assertEqual(net.gen_P_cost,gen_P_cost)
                 self.assertNotEqual(method_ipopt.results['net properties']['gen_P_cost'],gen_P_cost)
                 x1 = method_ipopt.get_results()['primal variables']
+                i1 = method_ipopt.get_results()['iterations']
                 p1 = method_ipopt.get_results()['net properties']['gen_P_cost']
-                print("%s  %s" %('IpoptOPF',case.split('/')[-1]))
+                print("%s\t%s\t%d" %(case.split('/')[-1],'ipopt',i1))
             except ImportError:
                 has_ipopt = False
-            
+
+            # INLP
+            net.update_properties()
+            gen_P_cost = net.gen_P_cost
+            method_inlp.solve(net)
+            self.assertEqual(method_inlp.results['status'],'solved')
+            self.assertEqual(net.gen_P_cost,gen_P_cost)
+            self.assertNotEqual(method_inlp.results['net properties']['gen_P_cost'],gen_P_cost)
+            x2 = method_inlp.get_results()['primal variables']
+            i2 = method_inlp.get_results()['iterations']
+            p2 = method_inlp.get_results()['net properties']['gen_P_cost']
+            print("%s\t%s\t%d" %(case.split('/')[-1],'inlp',i2))
+
+            # AUGL
             net.update_properties()
             gen_P_cost = net.gen_P_cost
             method_augl.solve(net)
             self.assertEqual(method_augl.results['status'],'solved')
             self.assertEqual(net.gen_P_cost,gen_P_cost)
             self.assertNotEqual(method_augl.results['net properties']['gen_P_cost'],gen_P_cost)
-            x2 = method_augl.get_results()['primal variables']
-            i2 = method_augl.get_results()['iterations']
-            p2 = method_augl.get_results()['net properties']['gen_P_cost']
-            print("%s  %s %d" %('AugLOPF ',case.split('/')[-1],i2))
-            if has_ipopt:
-                error = 100*(p1-p2)/abs(p2)
-                self.assertLess(np.abs(error),eps)
-                self.assertNotEqual(p1,p2)
+            x3 = method_augl.get_results()['primal variables']
+            i3 = method_augl.get_results()['iterations']
+            p3 = method_augl.get_results()['net properties']['gen_P_cost']
+            print("%s\t%s\t%d" %(case.split('/')[-1],'augl',i3))
 
-    @unittest.skip("temporarily disabled")            
+            # Checks
+            if has_ipopt:
+                error = 100*(p1-p3)/abs(p3)
+                self.assertLess(np.abs(error),eps)
+                self.assertNotEqual(p1,p3)
+            error = 100*(p2-p3)/abs(p3)
+            self.assertLess(np.abs(error),eps)
+            self.assertNotEqual(p2,p3)
+
     def test_DCOPF(self):
 
         T = 2
 
         infcases = ['ieee25.raw']
 
-        skipcases = ['case1354pegase.mat','case2869pegase.mat',
-                     'case3375wp.mat','case9241pegase.mat']
+        skipcases = ['case1354.mat','case2869.mat',
+                     'case3375wp.mat','case9241.mat']
         
         method = gopt.power_flow.new_method('DCOPF')
 
@@ -193,8 +220,8 @@ class TestPowerFlow(unittest.TestCase):
             self.assertEqual(net.num_periods,T)
             
             method.set_parameters({'quiet':True, 
-                                   'thermal_factor': 0.93,
-                                   'tol': 1e-6})
+                                   'tol': 1e-6,
+                                   'thermal_limits': True})
 
             try:
                 net.update_properties()
@@ -223,16 +250,16 @@ class TestPowerFlow(unittest.TestCase):
             x = results['primal variables']
             lam0,nu0,mu0,pi0 = results['dual variables']
 
-            self.assertTupleEqual(x.shape,((net.num_branches+
-                                            net.num_buses-
+            self.assertTupleEqual(x.shape,((net.num_buses-
                                             net.get_num_slack_buses()+
                                             net.get_num_P_adjust_gens())*T,))
-            self.assertTupleEqual(x.shape,(net.num_vars+net.num_branches*T,))
-            self.assertTupleEqual(lam0.shape,((net.num_buses+net.num_branches)*T,))
+            self.assertTupleEqual(x.shape,(net.num_vars,))
+            self.assertTupleEqual(lam0.shape,(net.num_buses*T,))
             self.assertTrue(nu0.size == 0)
-            self.assertTupleEqual(mu0.shape,x.shape)
-            self.assertTupleEqual(pi0.shape,x.shape)
+            self.assertTupleEqual(mu0.shape,(net.num_vars+net.num_branches*T,))
+            self.assertTupleEqual(pi0.shape,(net.num_vars+net.num_branches*T,))
 
+            # Network update (vars and sensitivities)
             xx = x[:net.num_vars]
             for t in range(T):
                 for bus in net.buses:
@@ -249,19 +276,7 @@ class TestPowerFlow(unittest.TestCase):
                     self.assertEqual(branch.sens_P_u_bound[t],mu0[net.num_vars+branch.index+t*net.num_branches])
                     self.assertEqual(branch.sens_P_l_bound[t],pi0[net.num_vars+branch.index+t*net.num_branches])
 
-            # gen outage 
-            if net.get_num_P_adjust_gens() > 1:
-                cont = pf.Contingency(gens=[net.get_generator(0)])
-                cont.apply()
-                try:
-                    method.solve(net)
-                    self.assertEqual(method.results['status'],'solved')
-                except gopt.power_flow.PFmethodError:
-                    self.assertTrue(case.split('/')[-1] in infcases)
-                    self.assertEqual(method.results['status'],'error')
-                cont.clear()
-
-            # no thermal limits
+            # No thermal limits
             method.set_parameters({'thermal_limits':False})
             method.solve(net)
             self.assertEqual(method.results['status'],'solved')
@@ -270,24 +285,23 @@ class TestPowerFlow(unittest.TestCase):
             gen_P_cost1 = net.gen_P_cost
             load_P_util1 = net.load_P_util
             lam1,nu1,mu1,pi1 = results['dual variables']
-            if ((norm(mu0[net.num_vars:],np.inf) > 1e-3 or 
-                 norm(pi0[net.num_vars:],np.inf) > 1e-3) and (case.split('/')[-1] not in infcases)):
-                self.assertTrue(np.all(gen_P_cost1 <= gen_P_cost0))
-            self.assertLess(norm(mu1[net.num_vars:],np.inf),1e-6)
-            self.assertLess(norm(pi1[net.num_vars:],np.inf),1e-6)
-           
-            # elastic loads
+            self.assertTupleEqual(mu1.shape,(net.num_vars,))
+            self.assertTupleEqual(pi1.shape,(net.num_vars,))
+
+            # Elastic loads
             for load in net.loads:
-                load.P_max = load.P[0]+1.
-                load.P_min = load.P[0]-1.
+                load.P_max = load.P+1.
+                load.P_min = load.P-1.
+            self.assertEqual(net.get_num_P_adjust_loads(),net.num_loads)
             for load in net.loads:
                 self.assertFalse(load.has_flags('variable','active power'))
                 self.assertFalse(load.has_flags('bounded','active power'))
+                self.assertTrue(np.all(load.P_min < load.P_max))
             method.solve(net)
             for load in net.loads:
                 self.assertTrue(load.has_flags('variable','active power'))
                 self.assertTrue(load.has_flags('bounded','active power'))
-                self.assertEqual(method.results['status'],'solved')
+            self.assertEqual(method.results['status'],'solved')
             results = method.get_results()
             method.update_network(net)
             self.assertTrue(np.all(net.gen_P_cost-net.load_P_util < gen_P_cost1-load_P_util1))
@@ -295,13 +309,12 @@ class TestPowerFlow(unittest.TestCase):
             x = results['primal variables']
             lam2,nu2,mu2,pi2 = results['dual variables']
 
-            self.assertTupleEqual(x.shape,((net.num_branches+
-                                            net.get_num_P_adjust_loads()+
+            self.assertTupleEqual(x.shape,((net.get_num_P_adjust_loads()+
                                             net.num_buses-
                                             net.get_num_slack_buses()+
                                             net.get_num_P_adjust_gens())*net.num_periods,))
-            self.assertTupleEqual(x.shape,(net.num_vars+net.num_branches*net.num_periods,))
-            self.assertTupleEqual(lam2.shape,((net.num_buses+net.num_branches)*net.num_periods,))
+            self.assertTupleEqual(x.shape,(net.num_vars,))
+            self.assertTupleEqual(lam2.shape,(net.num_buses*net.num_periods,))
             self.assertTrue(nu2.size == 0)
             self.assertTupleEqual(mu2.shape,x.shape)
             self.assertTupleEqual(pi2.shape,x.shape)
@@ -323,9 +336,6 @@ class TestPowerFlow(unittest.TestCase):
                         self.assertEqual(load.P[t],xx[load.index_P[t]])
                         self.assertEqual(load.sens_P_u_bound[t],mu2[load.index_P[t]])
                         self.assertEqual(load.sens_P_l_bound[t],pi2[load.index_P[t]])
-                for branch in net.branches:
-                    self.assertEqual(branch.sens_P_u_bound[t],mu2[net.num_vars+branch.index+t*net.num_branches])
-                    self.assertEqual(branch.sens_P_l_bound[t],pi2[net.num_vars+branch.index+t*net.num_branches])
                      
     def tearDown(self):